refactor(#1773): extract read_compressed_offset_window into its own module
Keep the already-over-threshold data_access/mod.rs from growing (campsite rule,
epic #1116): the new CRC-validated compressed offset-window helper moves to
data_access/compressed_offset.rs. mod.rs shrinks 888 -> 839.
